+++ src/main/routes/features/claim/claimantDoBController.ts
@@ -1,11 +1,11 @@
 import express from 'express';
-import {CLAIMANT_DOB_URL, CLAIMANT_PHONE_NUMBER_URL} from '../../../urls';
-import {GenericForm} from '../../../../common/form/models/genericForm';
-import {ClaimantDoB} from '../../../../common/form/models/claim/claimantDoB';
-import {Claim} from '../../../../common/models/claim';
-import {getCaseDataFromStore, saveDraftClaim} from '../../../../modules/draft-store/draftStoreService';
-import {AppRequest} from '../../../../common/models/AppRequest';
-import {Respondent} from '../../../../common/models/respondent';
+import {CLAIMANT_DOB_URL, CLAIMANT_PHONE_NUMBER_URL} from '../../urls';
+import {GenericForm} from '../../../common/form/models/genericForm';
+import {ClaimantDoB} from '../../../common/form/models/claim/claimantDoB';
+import {Claim} from '../../../common/models/claim';
+import {getCaseDataFromStore, saveDraftClaim} from '../../../modules/draft-store/draftStoreService';
+import {AppRequest} from '../../../common/models/AppRequest';
+import {Respondent} from '../../../common/models/respondent';
 
 const claimantDoBController = express.Router();
 const claimantDoBViewPath = 'features/response/citizenDob/citizen-dob';
