+++ src/main/routes/features/claim/claimantPartyTypeController.ts
@@ -1,9 +1,9 @@
 import express from 'express';
-import {CLAIMANT_PARTY_TYPE_SELECTION_URL} from '../../../urls';
-import {GenericForm} from '../../../../common/form/models/genericForm';
-import {PartyTypeSelection} from '../../../../common/form/models/claim/partyTypeSelection';
-import {ClaimantOrDefendant} from '../../../../common/models/partyType';
-import {redirectToPage} from '../../../../services/features/claim/partyTypeService';
+import {CLAIMANT_PARTY_TYPE_SELECTION_URL} from '../../urls';
+import {GenericForm} from '../../../common/form/models/genericForm';
+import {PartyTypeSelection} from '../../../common/form/models/claim/partyTypeSelection';
+import {ClaimantOrDefendant} from '../../../common/models/partyType';
+import {redirectToPage} from '../../../services/features/claim/partyTypeService';
 
 const claimantPartyTypeViewPath = 'features/public/claim/claimant-party-type';
 const claimantPartyTypeController = express.Router();
